pyproject.toml
[tool.poetry]
name = "chirper"
version = "0.1.0"
description = "Chirper API"
authors = ["Lucas Vizentin <Lucas.Vizentin@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.1"
factory-boy = "^3.3.0"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
